requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.4

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0

# NLP & Embeddings
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
tiktoken>=0.5.0

# LLM APIs
anthropic>=0.18.0
openai>=1.0.0

# Image Processing
pillow>=10.0.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web App
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.14.0

# Geocoding
geopy>=2.4.0

# Data Download
requests>=2.31.0
sodapy>=2.2.0

# Interpretability
shap>=0.43.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0
pyyaml>=6.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
